+++ backend/src/main/java/com/miu/book/BookCopy.java
@@ -1,0 +1,12 @@
+package com.miu.book;
+
+import com.miu.util.HelperUtil;
+
+public class BookCopy{
+    private String bookCopyId;
+    private boolean isAvailable;
+    public BookCopy(){
+        this.bookCopyId = HelperUtil.generateID();
+        this.isAvailable = true;
+    }
+}
