+++ backend/src/main/java/com/miu/book/Publication.java
@@ -1,0 +1,5 @@
+package com.miu.book;
+
+public interface Publication {
+
+}
