+++ backend/src/main/java/com/miu/dao/UserDao.java
@@ -1,6 +1,6 @@
 package com.miu.dao;
 
-import com.miu.User;
+import com.miu.person.User;
 import com.miu.dataStorage.DataStorage;
 
 import java.io.Serializable;
