+++ backend/src/main/java/com/miu/person/Address.java
@@ -1,4 +1,4 @@
-package com.miu;
+package com.miu.person;
 
 import java.io.Serializable;
