+++ backend/src/main/java/com/miu/person/Admin.java
@@ -1,4 +1,4 @@
-package com.miu;
+package com.miu.person;
 
 public class Admin extends UserRole {
     public Admin(String roleName) {
