+++ backend/src/main/java/com/miu/person/Librarian.java
@@ -1,6 +1,6 @@
-package com.miu;
+package com.miu.person;
 
-public class Librarian extends UserRole{
+public class Librarian extends UserRole {
     public Librarian(String roleName) {
         super(roleName);
     }
